refactor: move files to generic folder
