+++ src/generic/config/database/errors/DatabaseConnectionError.ts
