+++ src/index.ts
@@ -1,9 +1,9 @@
 import express, { Express } from "express";
-import defaultMiddlewareConfig from "./middlewares/defaultMiddleware.config";
-import configRoutes from "./middlewares/configRoutes";
-import environment from "./config/environment";
-import TypeOrmConnection from "./config/database/TypeOrmConnection";
-import postgresDataSource from "./config/database/datasources/postgresDataSource";
+import defaultMiddlewareConfig from "./generic/middlewares/defaultMiddleware.config";
+import configRoutes from "./generic/middlewares/configRoutes";
+import environment from "./generic/config/environment";
+import TypeOrmConnection from "./generic/config/database/TypeOrmConnection";
+import postgresDataSource from "./generic/config/database/datasources/postgresDataSource";
 
 const server: Express = express();
 defaultMiddlewareConfig(server);
