--- src/views/view-test.ejs
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt">
-<head>
-	<meta charset="UTF-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<title><%= title %></title>
-</head>
-<body>
-	<h1><%= message %></h1>
-	<img src="<%= imagePath %>" alt="coffee cat" style="width: 300px; height: auto;">
-</body>
-</html>
